+++ Source/RSP/Profiling.cpp
@@ -25,6 +25,7 @@
  */
 
 #include <windows.h>
+#include <shellapi.h>
 extern "C" {
 #include "profiling.h"
 }
@@ -129,8 +130,8 @@ public:
 			//sort the list with a basic bubble sort
 			if (ItemList.size() > 0)
 			{
-				for (int OuterPass = 0; OuterPass < (ItemList.size() - 1); OuterPass++ ) {
-					for (int InnerPass = 0; InnerPass < (ItemList.size() - 1); InnerPass++ ) {
+				for (size_t OuterPass = 0; OuterPass < (ItemList.size() - 1); OuterPass++ ) {
+					for (size_t InnerPass = 0; InnerPass < (ItemList.size() - 1); InnerPass++ ) {
 						if (ItemList[InnerPass]->second < ItemList[InnerPass + 1]->second) {
 							PROFILE_VALUE * TempPtr = ItemList[InnerPass];
 							ItemList[InnerPass] = ItemList[InnerPass + 1];
@@ -147,9 +148,9 @@ public:
 				{Timer_RDP_Running,   "RDP: Running"},
 			};
 			
-			for (int count =0; count < ItemList.size(); count++ ) {
+			for (size_t count = 0; count < ItemList.size(); count++ ) {
 				char Buffer[255];
-				float CpuUsage = ((double)ItemList[count]->second / (double)TotalTime) * 100;
+				float CpuUsage = (float)(((double)ItemList[count]->second / (double)TotalTime) * 100);
 				if (CpuUsage <= 0.2) { continue; }
 				sprintf(Buffer,"Func 0x%08X",ItemList[count]->first);
 				for (int NameID = 0; NameID < (sizeof(TimerNames) / sizeof(TIMER_NAME)); NameID++) {
